fix overflow error and added location title to pull down screen.

+++ lib/Models/WeatherModel.dart
@@ -89,7 +89,7 @@ class Clouds
 class Wind 
 {
     final double speed;
-    final double deg;
+    final int deg;
 
     Wind({
         this.speed,
